refactor: rename component and replace folders
