+++ src/features/auth/assets/images/EmailVerification/confirm.png
