+++ src/features/auth/assets/images/EmailVerification/expired.png
