+++ src/features/auth/ui/emailVerification/EmailVerification.module.scss
