+++ src/features/auth/ui/emailVerification/index.ts
@@ -1,0 +1,1 @@
+export { EmailVerification } from './EmailVerification'
