+++ src/shared/ui/index.ts
@@ -5,7 +5,6 @@ export { RadioGroup, RadioGroupItem, RadioGroupRoot } from './radioGroup'
 export { Sidebar } from './sidebar'
 export { Typography } from './typography'
 export { Modal } from './modal'
-export {InfoBlock } from './infoBlock'
 export {TextField } from './textField'
 export { DatePicker } from './datePicker'
 export {SelectBox, SelectItem } from './selectBox'
